重构游戏架构并新增基础实体系统
本次重大重构引入了全新的基础实体系统。新增了 Entity、Item、NPC 和 MessageBox 等基类，并将 Player 和 Boss 重构为使用新的架构。引入了 Item_Inventory 以及物品工厂/注册系统。事件绑定更新为支持世界坐标，改进了区域和按钮的控制逻辑，并在类型与 API 上做了多处优化，以增强一致性和可扩展性。移除了旧的物品代码，同时更新了 CMake 工程结构与命名规则。

+++ include/game/test/boss_a.hpp
@@ -219,8 +219,8 @@ namespace game {
             std::for_each(
                 std::execution::par_unseq, abc.begin(), abc.end(),
                 [this](auto &a) {
-                    a.setPosition(getX() + 60 * angle.getCos(), getY() + 60 * angle.getSin());
-                    angle.revolve(2 * sfui::PI / static_cast<float>(abc.size()));
+                    a.setPosition(getX() + 60 * static_cast<float>(angle.getCos()), getY() + 60 * static_cast<float>(angle.getSin()));
+                    angle.revolve(static_cast<float>(2 * sfui::PI )/ static_cast<float>(abc.size()));
                 });
 
             if (distance(player.getX(), player.getY(), getX(), getY()) > 1000) {
